Release

+++ scripts/local/setup.sh
@@ -12,23 +12,30 @@ if [[ ! -f ".lando.local.yml" ]]; then
   cp .lando.local.example.yml .lando.local.yml
 fi
 
+# YALESITES_BUILD_TOKEN is needed for authentication to Github.
+if [[ -z "$YALESITES_BUILD_TOKEN" ]]; then
+  echo "The YALESITES_BUILD_TOKEN variable must be set before setup can continue."
+  exit 1
+fi
+
 # Start lando and create containers.
 lando start
 
-# Install packages and pull down latest database and files.
+# Install packages and install Drupal using yalesites_profile.
 npm install
-npm run build-with-assets
+npm run build-with-install
 
 # Configure Composer to use source packaged versions.
 lando composer config --global 'preferred-install.yalesites-org/*' source
 
 # Manually remove the originally downloaded dist packed version.
-lando ssh -c "rm -rf web/profiles/contrib/yalesites_profile"
 lando ssh -c "rm -rf web/themes/contrib/atomic"
 
 # Use Composer to download the new version of the Yale projects.
-lando composer update yalesites_profile
 lando composer update atomic
+
+# Setup npm linked packages for theme dependencies
+npm run local:theme-link
 
 # Create a login link.
 lando drush uli
